refactor(mse-controller): cleanup typing to improve type safety
